sf_web_server/requirements/dev.txt
asgiref==3.5.2
cfgv==3.3.1
distlib==0.3.6
Django==4.1.2
django-extensions==3.2.1
djangorestframework==3.14.0
filelock==3.8.0
flake8==5.0.4
identify==2.5.8
mccabe==0.7.0
nodeenv==1.7.0
platformdirs==2.5.3
pre-commit==2.20.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.5
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
virtualenv==20.16.6
